Move identifier setters out of audio_analysis into metadata controller
