+++ music_assistant/controllers/metadata.py
@@ -373,6 +373,98 @@ class MetaDataController(CoreController):
                 )
         return item
 
+    async def set_track_identifiers(
+        self,
+        item_id: str,
+        provider_instance_id_or_domain: str,
+        mbid: str | None = None,
+        acoustid: str | None = None,
+        isrcs: list[str] | None = None,
+        media_type: MediaType = MediaType.TRACK,
+    ) -> None:
+        """
+        Persist MBID / AcoustID / ISRCs onto the library track row.
+
+        :param item_id: Provider-native track ID.
+        :param provider_instance_id_or_domain: Music provider instance ID or domain.
+        :param mbid: MusicBrainz recording ID.
+        :param acoustid: AcoustID UUID.
+        :param isrcs: ISRC codes.
+        :param media_type: Media type of the item.
+        """
+        # MBID is filled only when empty; AcoustID/ISRCs are appended via
+        # external_ids without clobbering tag-sourced values.
+        if media_type != MediaType.TRACK:
+            return
+        if not mbid and not acoustid and not isrcs:
+            return
+        try:
+            track = await self.mass.music.tracks.get_library_item_by_prov_id(
+                item_id, provider_instance_id_or_domain
+            )
+        except MusicAssistantError as err:
+            self.logger.debug(
+                "set_track_identifiers: failed to load library track %s/%s: %s",
+                provider_instance_id_or_domain,
+                item_id,
+                err,
+            )
+            return
+        if track is None:
+            return
+
+        changed = False
+        if mbid and not track.mbid:
+            track.mbid = mbid
+            changed = True
+        if acoustid and not any(
+            ext_id[0] == ExternalID.ACOUSTID and ext_id[1] == acoustid
+            for ext_id in track.external_ids
+        ):
+            track.add_external_id(ExternalID.ACOUSTID, acoustid)
+            changed = True
+        for isrc in isrcs or ():
+            if isrc:
+                track.add_external_id(ExternalID.ISRC, isrc)
+                changed = True
+        if not changed:
+            return
+
+        await self.mass.music.tracks.update_item_in_library(int(track.item_id), track)
+
+    async def set_album_release_group(
+        self,
+        album_item_id: int,
+        release_group_mbid: str,
+    ) -> None:
+        """
+        Persist a MusicBrainz release-group ID on a library album, idempotently.
+
+        :param album_item_id: Library album item_id (database id).
+        :param release_group_mbid: MusicBrainz release-group UUID to set.
+        """
+        if not release_group_mbid:
+            return
+        try:
+            album = await self.mass.music.albums.get_library_item(album_item_id)
+        except MusicAssistantError as err:
+            self.logger.debug(
+                "set_album_release_group: cannot load album %s: %s", album_item_id, err
+            )
+            return
+        # Refuse to overwrite — keeps tag-sourced or already-enriched IDs authoritative.
+        if album.get_external_id(ExternalID.MB_RELEASEGROUP):
+            self.logger.debug(
+                "set_album_release_group: album %s already has MB_RELEASEGROUP — keeping",
+                album_item_id,
+            )
+            return
+        album.add_external_id(ExternalID.MB_RELEASEGROUP, release_group_mbid)
+        await self.mass.music.albums.update_item_in_library(album_item_id, album)
+        self.logger.debug(
+            "set_album_release_group: wrote %s onto album %s", release_group_mbid, album_item_id
+        )
+
     def schedule_update_metadata(self, item: MediaItemType) -> None:
         """Schedule metadata update for given MediaItem."""
         if item.provider != "library":
