+++ tests/providers/acoustid_lookup/test_provider.py
@@ -36,11 +36,9 @@ def _make_provider(
     mass.get_provider = MagicMock(return_value=None)
     mass.streams.audio_analysis.get_audio_analysis_version = AsyncMock(return_value=None)
     mass.streams.audio_analysis.set_audio_analysis = AsyncMock()
-    mass.streams.audio_analysis.set_track_identifiers = AsyncMock()
-    mass.streams.audio_analysis.set_album_release_group = AsyncMock()
-    mass.streams.audio_analysis.get_acoustid_extra_data_for_album_tracks = AsyncMock(
-        return_value=[]
-    )
+    mass.metadata.set_track_identifiers = AsyncMock()
+    mass.metadata.set_album_release_group = AsyncMock()
+    mass.streams.audio_analysis.get_extra_data_for_album_tracks = AsyncMock(return_value=[])
     mass.music.tracks.get_library_item_by_prov_id = AsyncMock(return_value=None)
     mass.music.albums.get_library_item = AsyncMock(return_value=None)
     mass.music.albums.get_library_album_tracks = AsyncMock(return_value=[])
@@ -194,7 +192,7 @@ def _wire_album_for_consensus(
     )
     cast(
         "MagicMock", provider.mass.streams.audio_analysis
-    ).get_acoustid_extra_data_for_album_tracks = AsyncMock(return_value=extras)
+    ).get_extra_data_for_album_tracks = AsyncMock(return_value=extras)
 
 
 def _rg(
@@ -377,9 +375,7 @@ async def test_post_analysis_persists_and_writes_tags(
 
     await provider.post_analysis(streamdetails, analysis)
 
-    cast(
-        "AsyncMock", provider.mass.streams.audio_analysis.set_track_identifiers
-    ).assert_awaited_once_with(
+    cast("AsyncMock", provider.mass.metadata.set_track_identifiers).assert_awaited_once_with(
         item_id=streamdetails.item_id,
         provider_instance_id_or_domain=streamdetails.provider,
         mbid="mbid-x",
@@ -458,7 +454,7 @@ async def test_post_analysis_mb_enrichment(
         AudioAnalysisData(extra_data={"mbid": "mbid-x", "acoustid": "acoustid-x"}),
     )
 
-    set_ids = cast("AsyncMock", provider.mass.streams.audio_analysis.set_track_identifiers)
+    set_ids = cast("AsyncMock", provider.mass.metadata.set_track_identifiers)
     assert set_ids.await_args is not None
     assert set_ids.await_args.kwargs["isrcs"] == expected_isrcs
     if expected_isrcs:
@@ -488,9 +484,7 @@ async def test_post_analysis_consensus_silent_failure(monkeypatch: pytest.Monkey
         ),
     )
 
-    cast(
-        "AsyncMock", provider.mass.streams.audio_analysis.set_track_identifiers
-    ).assert_awaited_once()
+    cast("AsyncMock", provider.mass.metadata.set_track_identifiers).assert_awaited_once()
 
 
 # ---------------------------------------------------------------------------
@@ -530,9 +524,9 @@ async def test_consensus_writes_release_group(
 
     await provider._maybe_set_album_release_group(_make_streamdetails())
 
-    cast(
-        "AsyncMock", provider.mass.streams.audio_analysis.set_album_release_group
-    ).assert_awaited_once_with(42, "rg-st")
+    cast("AsyncMock", provider.mass.metadata.set_album_release_group).assert_awaited_once_with(
+        42, "rg-st"
+    )
 
 
 @pytest.mark.asyncio
@@ -591,9 +585,7 @@ async def test_consensus_abstains(
 
     await provider._maybe_set_album_release_group(_make_streamdetails())
 
-    cast(
-        "AsyncMock", provider.mass.streams.audio_analysis.set_album_release_group
-    ).assert_not_awaited()
+    cast("AsyncMock", provider.mass.metadata.set_album_release_group).assert_not_awaited()
 
 
 # ---------------------------------------------------------------------------
